Vendor jeat and some bundle files via jeep
The transition to using jeep vendoring was completed.

With the git-subrepo-tracked jeat directory out of the way with the
last commit, jeep was used to vendor jeat as well as some bundle/
files.

Steps followed included:

* Creating an appropriate info.jdn with a :vendored section
  so that jeat and some-bundle-bits were referenced along
  with specifying which files to copy as well as where to
  copy those files to.

* Jeep was invoked to carry out the copying.  The invocation
  was: jeep prep vendor

With the steps completed, testing, which had been disabled with the
last commit, was re-enabled.